chain/symbol.go: Make Precision() return int64
Should be enough, alotho it can still overflow because Decimals() can
return 0-255 resulting in Precision = 10^255

+++ chain/asset_test.go
@@ -47,14 +47,14 @@ func TestAsset(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, asset3.Value, int64(10000))
 	assert.Equal(t, asset3.Decimals(), 4)
-	assert.Equal(t, asset3.Precision(), 10000)
+	assert.Equal(t, asset3.Precision(), int64(10000))
 	assert.Equal(t, asset3.String(), "1.0000 EOS")
 
 	asset4, err := chain.NewAssetFromString("34.0303 EOS")
 	assert.NoError(t, err)
 	assert.Equal(t, asset4.Value, int64(340303))
 	assert.Equal(t, asset4.Decimals(), 4)
-	assert.Equal(t, asset4.Precision(), 10000)
+	assert.Equal(t, asset4.Precision(), int64(10000))
 	assert.Equal(t, asset4.String(), "34.0303 EOS")
 
 	asset5, err := chain.NewAssetFromString("1.123456789012345678 ABCDEFG")
